style: fiber 错误处理

+++ internal/status/errors.go
@@ -5,7 +5,7 @@ import "fmt"
 // SysError 服务业务错误
 type SysError struct {
 	// 错误码
-	Code int8 `json:"code"`
+	Code int `json:"code"`
 	// 错误消息
 	Message string `json:"message"`
 }
@@ -15,6 +15,6 @@ func (be *SysError) Error() string {
 }
 
 // 构建新的状态
-func ofErr(code int8, msg string) *SysError {
+func ofErr(code int, msg string) *SysError {
 	return &SysError{code, msg}
 }
